Code review (rev 25): Harden WAV parser, fix signed/unsigned mismatch, make client game counter atomic, add PID file error handling
- androidsoundeffect: Validate all WAV header reads (check return values),
  use qint32 instead of off_t for data chunk length to match WAV spec,
  reject negative data lengths that could cause huge allocations
- cleanerserver: Fix signed/unsigned truncation in constant-time
  comparison loop (qsizetype vs int) using auto/decltype
- clientthread: Make m_curGameNum atomic to match server-side pattern
  and prevent potential data race
- pokerth_server: Use error_code overload for PID file removal and
  log failure instead of silently ignoring errors

+++ src/gui/qt/sound/androidsoundeffect.h
@@ -58,7 +58,7 @@ public slots:
 
 private:
 	std::vector<char> mBuffer;
-	off_t mLength = 0;
+	qint32 mLength = 0;
 
 	QString mPath;
